refactor: consolidate stop reference building logic for improved clarity and reuse
